+++ pkg/js/dev/myams-core-dev.js
@@ -550,7 +550,7 @@ function getScript(url) {
   return new Promise((resolve, reject) => {
     const defaults = {
       dataType: 'script',
-      url: getSource(url),
+      url: MyAMS.core.getSource(url),
       cache: MyAMS.env.devmode,
       async: true
     };
@@ -665,15 +665,17 @@ function switchIcon(element, fromClass, toClass) {
 }
 
 /**
- * MyAMS base environment
+ * MyAMS base environment getter
  *
- * @type {{
- *     bundle: boolean,
- *     devmode: boolean,
- *     devext: string,
- *     extext: string,
- *     theme: string,
- *     baseURL: string
+ * @type {Object}
+ *
+ * Returns an object with the following attributes matching MyAMS environment:
+ * - bundle: boolean; true if MyAMS is published using modules bundle
+ * - devmode: boolean; true if MyAMS is published in development mode
+ * - devext: string: extension used in development mode
+ * - extext: string: extension used for external extensions
+ * - theme: string: current MyAMS theme name
+ * - baseURL: string: base MyAMS URL
  * }}
  */
 function getEnv($) {
@@ -695,10 +697,8 @@ function getEnv($) {
  * MyAMS theme getter
  */
 function getTheme() {
-  let theme;
-  if (MyAMS.env.bundle) {
-    theme = MyAMS.theme;
-  } else {
+  let theme = MyAMS.theme;
+  if (!theme) {
     const css = $('link[href*="/myams.css"], link[href*="/emerald.css"], link[href*="/darkmode.css"]');
     theme = css.length > 0 ? /.*\/([a-z]+).css/.exec(css.attr('href'))[1] : 'unknown';
   }
@@ -722,23 +722,25 @@ function getDOM() {
 /**
  * MyAMS default configuration
  *
- * @type {Object}:
- *      modules: array of loaded extension modules
- *      ajaxNav: true if AJAX navigation is enabled
- *      enableFastclick: true is "smart-click" extension is to be activated on mobile devices
- *      menuSpeed: menu speed, in miliseconds
- *      initPage: dotted name of MyAMS global init function
- *      initContent: dotted name of MyAMS content init function
- *      alertContainerCLass: class of MyAMS alerts container
- *      safeMethods: HTTP methods which can be used without CSRF cookie verification
- *      csrfCookieName: CSRF cookie name
- *      csrfHeaderName: CSRF header name
- *      enableTooltips: global tooltips enable flag
- *      enableHtmlTooltips: allow HTML code in tooltips
- *      warnOnFormChange: flag to specify if form changes should be warned
- *      formChangeCallback: global form change callback
- *      isMobile: boolean, true if device is detected as mobile
- *      device: string: 'mobile' or 'desktop'
+ * @type {Object}
+ *
+ * Returns an object matching current MyAMS configuration:
+ * - modules: array of loaded extension modules
+ * - ajaxNav: true if AJAX navigation is enabled
+ * - enableFastclick: true is "smart-click" extension is to be activated on mobile devices
+ * - menuSpeed: menu speed, in miliseconds
+ * - initPage: dotted name of MyAMS global init function
+ * - initContent: dotted name of MyAMS content init function
+ * - alertContainerCLass: class of MyAMS alerts container
+ * - safeMethods: HTTP methods which can be used without CSRF cookie verification
+ * - csrfCookieName: CSRF cookie name
+ * - csrfHeaderName: CSRF header name
+ * - enableTooltips: global tooltips enable flag
+ * - enableHtmlTooltips: allow HTML code in tooltips
+ * - warnOnFormChange: flag to specify if form changes should be warned
+ * - formChangeCallback: global form change callback
+ * - isMobile: boolean, true if device is detected as mobile
+ * - device: string: 'mobile' or 'desktop'
  */
 const isMobile = /iphone|ipad|ipod|android|blackberry|mini|windows\sce|palm/i.test(navigator.userAgent.toLowerCase()),
   config = {
@@ -1270,7 +1272,7 @@ if (html.data('ams-init') !== false) {
   (0,_ext_base__WEBPACK_IMPORTED_MODULE_0__.init)(_ext_base__WEBPACK_IMPORTED_MODULE_0__["default"].$);
 }
 
-/** Version: 1.15.0  */
+/** Version: 1.15.2  */
 }();
 /******/ })()
 ;